Back: suite uplaods (voir TODO) + gestion erreurs à finir

+++ backend/src/controllers/post.services.ts
@@ -14,25 +14,20 @@ export const serviceCreatePost = async (
 ): Promise<QueryError | boolean | unknown> => {
   const { content, title } = req.body;
   const postImgUrl = createPostImgUrl(req);
-  try {
-    const userId = await getUserId(req);
-    return new Promise((resolve, reject) => {
-      const reqCreatePost: string = `INSERT INTO posts (p_uid, p_content, p_post_img_url, p_creation_date, p_title, p_modification_date, p_fk_user_id) VALUES (UUID(), "${content}", "${postImgUrl}", NOW(), "${title}", NULL, "${userId}") `;
-      db.query(reqCreatePost, (err: QueryError) => {
-        err ? reject(err) : resolve(true);
-      });
+  const userId = await getUserId(req);
+  return new Promise((resolve, reject) => {
+    const reqCreatePost: string = `INSERT INTO posts (p_uid, p_content, p_post_img_url, p_creation_date, p_title, p_modification_date, p_fk_user_id) VALUES (UUID(), "${content}", "${postImgUrl}", NOW(), "${title}", NULL, "${userId}") `;
+    db.query(reqCreatePost, (err: QueryError) => {
+      err ? (console.log(err), reject(Error("query error"))) : resolve(true);
     });
-  } catch (err) {
-    console.log(err);
-    return err;
-  }
+  });
 };
 
 export const serviceGetAllPosts = (): Promise<QueryError | RowDataPacket[]> => {
   return new Promise((resolve, reject) => {
     const reqGetAllPosts: string = `SELECT p_uid, p_content, p_post_img_url, p_creation_date, p_title, p_modification_date, u_username FROM posts INNER JOIN users ON p_fk_user_id =  u_id`;
     db.query(reqGetAllPosts, (err: QueryError, rows: RowDataPacket[]) => {
-      err ? reject(err) : resolve(rows);
+      err ? (console.log(err), reject(Error("query error"))) : resolve(rows);
     });
   });
 };
@@ -43,7 +38,7 @@ export const serviceGetOnePost = (
   return new Promise((resolve, reject) => {
     const reqGetOnePost: string = `SELECT p_uid, p_content, p_post_img_url, p_creation_date, p_title, p_modification_date, u_username FROM posts INNER JOIN users ON p_fk_user_id = u_id WHERE p_uid = "${req.params.id}"`;
     db.query(reqGetOnePost, (err: QueryError, rows: RowDataPacket[0]) => {
-      err ? reject(err) : resolve(rows[0]);
+      err ? (console.log(err), reject(Error("query error"))) : resolve(rows[0]);
     });
   });
 };
@@ -54,7 +49,7 @@ export const serviceGetPostsByAuthor = (
   return new Promise((resolve, reject) => {
     const reqGetPostsByAuthor: string = `SELECT p_uid, p_content, p_post_img_url, p_creation_date, p_title, p_modification_date, u_username FROM users INNER JOIN posts ON u_id = p_fk_user_id WHERE u_uid = "${req.body.author}"`;
     db.query(reqGetPostsByAuthor, (err: QueryError, rows: RowDataPacket[]) => {
-      err ? reject(err) : resolve(rows);
+      err ? (console.log(err), reject(Error("query error"))) : resolve(rows);
     });
   });
 };
@@ -64,26 +59,19 @@ export const serviceUpdatePost = async (
 ): Promise<QueryError | boolean | unknown> => {
   const postId: string = req.params.id;
   const { content, title } = req.body;
-  try {
-    const postDatas = await checkIfPostExistAndGetDatas(postId);
-    console.log(postDatas);
-    const postOwner = postDatas.u_uid;
-    const oldPostImgUrl = postDatas.p_post_img_url;
-    console.log("Log oldPostImgUrl dans serviceUpdate: ", oldPostImgUrl);
-    const postImgUrl = setPostImgUrl(req, oldPostImgUrl);
-    return !postDatas
-      ? false
-      : postOwner === req.userUid
-      ? new Promise((resolve, reject) => {
-          const reqUpdatePost: string = `UPDATE posts SET p_content = '${content}', p_post_img_url = '${postImgUrl}',p_title = '${title}' WHERE p_uid = '${postId}'`;
-          db.query(reqUpdatePost, (err: QueryError) => {
-            err ? reject(err) : resolve(true);
-          });
-        })
-      : "Forbidden";
-  } catch (err) {
-    console.log(err);
-    return err;
+  const postDatas = await checkIfPostExistAndGetDatas(postId);
+  const postOwner = postDatas.u_uid;
+  const oldPostImgUrl = postDatas.p_post_img_url;
+  const postImgUrl = setPostImgUrl(req, oldPostImgUrl);
+  if (postOwner === req.userUid) {
+    return new Promise((resolve, reject) => {
+      const reqUpdatePost: string = `UPDATE posts SET p_content = '${content}', p_post_img_url = '${postImgUrl}',p_title = '${title}' WHERE p_uid = '${postId}'`;
+      db.query(reqUpdatePost, (err: QueryError) => {
+        err ? (console.log(err), reject(Error("query error"))) : resolve(true);
+      });
+    });
+  } else {
+    throw Error("forbidden");
   }
 };
 
@@ -91,24 +79,18 @@ export const serviceDeletePost = async (
   req: Request | any
 ): Promise<QueryError | boolean | unknown> => {
   const postId: string = req.params.id;
-  try {
-    const postDatas = await checkIfPostExistAndGetDatas(postId);
-    const postOwner = postDatas.u_uid;
-    const postImgUrl = postDatas.p_post_img_url;
-    console.log("Log postImgUrl dans service delete: ", postImgUrl);
-    return !postDatas
-      ? false
-      : postOwner === req.userUid
-      ? new Promise((resolve, reject) => {
-          deleteOldPostImageOnServer(postImgUrl);
-          const reqDeletePost: string = `DELETE FROM posts WHERE p_uid = '${postId}'`;
-          db.query(reqDeletePost, (err: QueryError) => {
-            err ? reject(err) : resolve(true);
-          });
-        })
-      : "Forbidden";
-  } catch (err) {
-    console.log(err);
-    return err;
+  const postDatas = await checkIfPostExistAndGetDatas(postId);
+  const postOwner = postDatas.u_uid;
+  const postImgUrl = postDatas.p_post_img_url;
+  if (postOwner === req.userUid) {
+    return new Promise((resolve, reject) => {
+      deleteOldPostImageOnServer(postImgUrl);
+      const reqDeletePost: string = `DELETE FROM posts WHERE p_uid = '${postId}'`;
+      db.query(reqDeletePost, (err: QueryError) => {
+        err ? (console.log(err), reject(Error("query error"))) : resolve(true);
+      });
+    });
+  } else {
+    throw Error("forbidden");
   }
 };
